requirements.txt
black
django
djangorestframework
django-cors-headers
django-filter
django-rest-auth
ipdb
markdown